Made zimtohrli a .h-only library.
Tests, benchmarks, compare binary, go-bindings - are now all objects that just include zimtohrli.h.
