+++ cpp/zimt/compare.cc
@@ -29,7 +29,6 @@
 #include "absl/strings/string_view.h"
 #include "sndfile.h"
 #include "zimt/audio.h"
-#include "zimt/fourier_bank.h"
 #include "zimt/mos.h"
 #include "zimt/zimtohrli.h"
 
@@ -105,8 +104,7 @@ int Main(int argc, char* argv[]) {
   float file_a_max_abs_amplitude = 0;
   for (size_t channel_index = 0; channel_index < file_a->Info().channels;
        ++channel_index) {
-    EnergyAndMaxAbsAmplitude measurements =
-        Measure(file_a->Buffer()[channel_index]);
+    EnergyAndMaxAbsAmplitude measurements = Measure((*file_a)[channel_index]);
     file_a_max_abs_amplitude =
         std::max(file_a_max_abs_amplitude, measurements.max_abs_amplitude);
     file_a_measurements.push_back(measurements);
@@ -115,7 +113,7 @@ int Main(int argc, char* argv[]) {
   if (verbose) {
     PrintLoadFileInfo(path_a, file_a->Info(), file_a_measurements);
   }
-  size_t min_length = file_a->Buffer().num_frames;
+  size_t min_length = file_a->Info().frames;
 
   std::vector<AudioFile> file_b_vector;
   file_b_vector.reserve(path_b.size());
@@ -133,19 +131,20 @@ int Main(int argc, char* argv[]) {
     std::vector<EnergyAndMaxAbsAmplitude> measurements;
     for (size_t channel_index = 0; channel_index < file_b->Info().channels;
          ++channel_index) {
-      measurements.push_back(Measure(file_b->Buffer()[channel_index]));
+      measurements.push_back(Measure((*file_b)[channel_index]));
     }
     if (verbose) {
       PrintLoadFileInfo(file_b->Path(), file_b->Info(), measurements);
     }
     CHECK_EQ(file_a->Info().channels, file_b->Info().channels);
     CHECK_EQ(file_a->Info().samplerate, file_b->Info().samplerate);
-    min_length = std::min(min_length, file_b->Buffer().num_frames);
+    min_length =
+        std::min(min_length, static_cast<size_t>(file_b->Info().frames));
     for (size_t channel_index = 0; channel_index < file_b->Info().channels;
          ++channel_index) {
       const EnergyAndMaxAbsAmplitude new_energy_and_max_abs_amplitude =
           NormalizeAmplitude(file_a_max_abs_amplitude,
-                             file_b->Buffer()[channel_index]);
+                             (*file_b)[channel_index]);
       if (verbose) {
         std::cerr << "  Normalized channel " << channel_index << " energy = "
                   << new_energy_and_max_abs_amplitude.energy_db_fs
@@ -169,27 +168,27 @@ int Main(int argc, char* argv[]) {
   };
 
   const bool per_channel = absl::GetFlag(FLAGS_per_channel);
-  const size_t num_downscaled_samples_a = static_cast<size_t>(
-      std::ceil(static_cast<float>(file_a->Buffer().num_frames) *
-                z.perceptual_sample_rate / kSampleRate));
+  const size_t num_downscaled_samples_a =
+      static_cast<size_t>(std::ceil(static_cast<float>(file_a->Info().frames) *
+                                    z.perceptual_sample_rate / kSampleRate));
   std::vector<Spectrogram> file_a_spectrograms;
   for (size_t channel_index = 0; channel_index < file_a->Info().channels;
        ++channel_index) {
     Spectrogram spectrogram(num_downscaled_samples_a, kNumRotators);
-    z.Analyze(file_a->Buffer()[channel_index], spectrogram);
+    z.Analyze((*file_a)[channel_index], spectrogram);
     file_a_spectrograms.push_back(std::move(spectrogram));
   }
   for (int file_b_index = 0; file_b_index < file_b_vector.size();
        ++file_b_index) {
     const AudioFile& file_b = file_b_vector[file_b_index];
-    const size_t num_downscaled_samples_b = static_cast<size_t>(
-        std::ceil(static_cast<float>(file_b.Buffer().num_frames) *
-                  z.perceptual_sample_rate / kSampleRate));
+    const size_t num_downscaled_samples_b =
+        static_cast<size_t>(std::ceil(static_cast<float>(file_b.Info().frames) *
+                                      z.perceptual_sample_rate / kSampleRate));
     Spectrogram spectrogram_b(num_downscaled_samples_b, kNumRotators);
     float sum_of_squares = 0;
     for (size_t channel_index = 0; channel_index < file_a->Info().channels;
          ++channel_index) {
-      z.Analyze(file_b.Buffer()[channel_index], spectrogram_b);
+      z.Analyze(file_b[channel_index], spectrogram_b);
       const float distance =
           z.Distance(file_a_spectrograms[channel_index], spectrogram_b);
       if (per_channel) {
